Refactored crawler and moved to separate file
Added rewriteScript
Added rewritePath
Added parseElement
